requirements.txt
django
psycopg2
django-rest-framework
django-webpack-loader
django-datatables-view
invoke
django-extensions
factory_boy